requirements.txt
# devops
Cython 
wheel
python-box[all]~=7.0
loguru
# apps
streamlit
